Statistical and configuration changes
I transitioned from the previously utilized statistics library to utilizing the chart.js library alongside its Svelte add-on. This enhancement enabled the creation of both a chart and a pie chart within the primary statistics tab. On the server side, I implemented a load limit and introduced the capability to modify the data for the initially created user via the environment file. Furthermore, I relocated the statistics acquisition logic to the dedicated statistics file